safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/mozilla/versionchange-abort-expected.txt

CONSOLE MESSAGE: line 190: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
When a versionchange transaction is aborted, the version and newly created object stores should be reset

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "versionchange-abort.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
